finni/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["jeongmiri <1ive4eva0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
django-environ = "^0.11.2"
gunicorn = "^21.2.0"
mysqlclient = "^2.2.4"
django-redis = "^5.4.0"
requests = "^2.31.0"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
setuptools = "^69.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
